+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Adds create PR URL to branch strategy diagram
 - Allow to connect to Azure with Personal Access Token (only way when your are guest user on another tenant)
+- Fix issue when fetching PRs and builds for PRs in Azure DevOps
 
 ## [6.10.1] 2024-10-08